requirements.txt
# Enhanced AI Reverse Engineering System - Requirements
# Core dependencies for analyzing AI-generated videos and images with advanced algorithms

# Computer Vision and Image Processing  
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
imageio>=2.31.0
ffmpeg-python>=0.2.0

# Advanced Image Analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# AI and Machine Learning
groq>=0.4.0

# Web Interface
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0

# System Requirements
pathlib2>=2.3.7
argparse>=1.4.0
